feat: add new music
Fixes some weird looping issues when it started in the StartState, so
that's much better. I thought it would be helpful to commit the
production files so they're easy to access over time and all in the
repo. Seems sensible for now.

Includes some formatting changes from switching to VS Code.